chore: extract collections to :ui:collections
