+++ app/config/baseline.xml
@@ -15,10 +15,7 @@
     <ID>ComposableParamOrder:SeriesStatsScreen.kt$SeriesStatsScreen</ID>
     <ID>ComposableParamOrder:WorkDetailsScreen.kt$WorkDetailsScreen</ID>
     <ID>ComposableParamOrder:WorkStatsScreen.kt$WorkGroupStatsScreen</ID>
-    <ID>MaxLineLength:ArtistsByCollectionViewModel.kt$ArtistsByCollectionViewModel$)</ID>
     <ID>MaxLineLength:InstrumentDetailsScreen.kt$description</ID>
-    <ID>MaxLineLength:InstrumentsByCollectionViewModel.kt$InstrumentsByCollectionViewModel$)</ID>
-    <ID>MaxLineLength:RecordingsByCollectionViewModel.kt$RecordingsByCollectionViewModel$)</ID>
     <ID>MaxLineLength:RecordingsByWorkViewModel.kt$RecordingsByWorkViewModel$)</ID>
     <ID>ModifierWithoutDefault:ReleaseGroupsByArtistScreen.kt$modifier</ID>
     <ID>TooGenericExceptionCaught:TopLevelViewModel.kt$TopLevelViewModel$ex: Exception</ID>
