+++ app/src/androidTest/java/ly/david/mbjc/ui/collection/CollectionParameterizedTest.kt
@@ -23,8 +23,8 @@ import ly.david.data.room.collection.CollectionRoomModel
 import ly.david.mbjc.MainActivityTest
 import ly.david.mbjc.StringReferences
 import ly.david.mbjc.ui.TopLevelScaffold
-import ly.david.mbjc.ui.collections.CollectionListScaffold
-import ly.david.mbjc.ui.collections.CollectionScaffold
+import ly.david.ui.collections.CollectionListScaffold
+import ly.david.ui.collections.CollectionScaffold
 import ly.david.mbjc.ui.navigation.goToEntityScreen
 import ly.david.ui.common.topappbar.TopAppBarWithFilterTestTag
 import ly.david.ui.core.theme.PreviewTheme
