--- app/src/main/java/ly/david/mbjc/ui/collections/artists/ArtistsByCollectionScreen.kt
@@ -1,76 +1,0 @@
-package ly.david.mbjc.ui.collections.artists
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.collectAsLazyPagingItems
-import ly.david.data.domain.listitem.ArtistListItemModel
-import ly.david.data.getNameWithDisambiguation
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.ui.common.artist.ArtistListItem
-import ly.david.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.ui.common.paging.PagingLoadingAndErrorHandler
-import ly.david.ui.common.rememberFlowWithLifecycleStarted
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-internal fun ArtistsByCollectionScreen(
-    collectionId: String,
-    isRemote: Boolean,
-    filterText: String,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    onArtistClick: (entity: MusicBrainzEntity, String, String) -> Unit = { _, _, _ -> },
-    onDeleteFromCollection: (entityId: String, name: String) -> Unit = { _, _ -> },
-    viewModel: ArtistsByCollectionViewModel = hiltViewModel(),
-) {
-    val entity = MusicBrainzEntity.ARTIST
-    val lazyListState = rememberLazyListState()
-    val lazyPagingItems: LazyPagingItems<ArtistListItemModel> =
-        rememberFlowWithLifecycleStarted(viewModel.pagedEntities)
-            .collectAsLazyPagingItems()
-
-    LaunchedEffect(key1 = collectionId) {
-        viewModel.setRemote(isRemote)
-        viewModel.loadPagedEntities(collectionId)
-    }
-
-    LaunchedEffect(key1 = filterText) {
-        viewModel.updateQuery(filterText)
-    }
-
-    PagingLoadingAndErrorHandler(
-        modifier = modifier,
-        lazyListState = lazyListState,
-        lazyPagingItems = lazyPagingItems,
-        snackbarHostState = snackbarHostState
-    ) { listItemModel: ArtistListItemModel? ->
-
-        when (listItemModel) {
-            is ArtistListItemModel -> {
-                SwipeToDeleteListItem(
-                    content = {
-                        ArtistListItem(
-                            artist = listItemModel,
-                            modifier = Modifier.animateItemPlacement(),
-                        ) {
-                            onArtistClick(entity, id, getNameWithDisambiguation())
-                        }
-                    },
-                    onDelete = {
-                        onDeleteFromCollection(listItemModel.id, listItemModel.name)
-                    }
-                )
-            }
-
-            else -> {
-                // Do nothing.
-            }
-        }
-    }
-}
