--- app/src/main/java/ly/david/mbjc/ui/collections/events/EventsByCollectionScreen.kt
@@ -1,75 +1,0 @@
-package ly.david.mbjc.ui.collections.events
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.collectAsLazyPagingItems
-import ly.david.data.domain.listitem.EventListItemModel
-import ly.david.data.getNameWithDisambiguation
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.ui.common.event.EventListItem
-import ly.david.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.ui.common.paging.PagingLoadingAndErrorHandler
-import ly.david.ui.common.rememberFlowWithLifecycleStarted
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-internal fun EventsByCollectionScreen(
-    collectionId: String,
-    isRemote: Boolean,
-    filterText: String,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    onEventClick: (entity: MusicBrainzEntity, String, String) -> Unit = { _, _, _ -> },
-    onDeleteFromCollection: (entityId: String, name: String) -> Unit = { _, _ -> },
-    viewModel: EventsByCollectionViewModel = hiltViewModel(),
-) {
-    val entity = MusicBrainzEntity.EVENT
-    val lazyListState = rememberLazyListState()
-    val lazyPagingItems: LazyPagingItems<EventListItemModel> =
-        rememberFlowWithLifecycleStarted(viewModel.pagedEntities)
-            .collectAsLazyPagingItems()
-
-    LaunchedEffect(key1 = collectionId) {
-        viewModel.setRemote(isRemote)
-        viewModel.loadPagedEntities(collectionId)
-    }
-
-    LaunchedEffect(key1 = filterText) {
-        viewModel.updateQuery(filterText)
-    }
-
-    PagingLoadingAndErrorHandler(
-        modifier = modifier,
-        lazyListState = lazyListState,
-        lazyPagingItems = lazyPagingItems,
-        snackbarHostState = snackbarHostState
-    ) { listItemModel: EventListItemModel? ->
-        when (listItemModel) {
-            is EventListItemModel -> {
-                SwipeToDeleteListItem(
-                    content = {
-                        EventListItem(
-                            event = listItemModel,
-                            modifier = Modifier.animateItemPlacement(),
-                        ) {
-                            onEventClick(entity, id, getNameWithDisambiguation())
-                        }
-                    },
-                    onDelete = {
-                        onDeleteFromCollection(listItemModel.id, listItemModel.name)
-                    }
-                )
-            }
-
-            else -> {
-                // Do nothing.
-            }
-        }
-    }
-}
