--- app/src/main/java/ly/david/mbjc/ui/collections/places/PlacesByCollectionScreen.kt
@@ -1,75 +1,0 @@
-package ly.david.mbjc.ui.collections.places
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.collectAsLazyPagingItems
-import ly.david.data.domain.listitem.PlaceListItemModel
-import ly.david.data.getNameWithDisambiguation
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.ui.common.paging.PagingLoadingAndErrorHandler
-import ly.david.ui.common.place.PlaceListItem
-import ly.david.ui.common.rememberFlowWithLifecycleStarted
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-internal fun PlacesByCollectionScreen(
-    collectionId: String,
-    isRemote: Boolean,
-    filterText: String,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    onPlaceClick: (entity: MusicBrainzEntity, String, String) -> Unit = { _, _, _ -> },
-    onDeleteFromCollection: (entityId: String, name: String) -> Unit = { _, _ -> },
-    viewModel: PlacesByCollectionViewModel = hiltViewModel(),
-) {
-    val entity = MusicBrainzEntity.PLACE
-    val lazyListState = rememberLazyListState()
-    val lazyPagingItems: LazyPagingItems<PlaceListItemModel> =
-        rememberFlowWithLifecycleStarted(viewModel.pagedEntities)
-            .collectAsLazyPagingItems()
-
-    LaunchedEffect(key1 = collectionId) {
-        viewModel.setRemote(isRemote)
-        viewModel.loadPagedEntities(collectionId)
-    }
-
-    LaunchedEffect(key1 = filterText) {
-        viewModel.updateQuery(filterText)
-    }
-
-    PagingLoadingAndErrorHandler(
-        lazyPagingItems = lazyPagingItems,
-        modifier = modifier,
-        lazyListState = lazyListState,
-        snackbarHostState = snackbarHostState
-    ) { listItemModel: PlaceListItemModel? ->
-        when (listItemModel) {
-            is PlaceListItemModel -> {
-                SwipeToDeleteListItem(
-                    content = {
-                        PlaceListItem(
-                            place = listItemModel,
-                            modifier = Modifier.animateItemPlacement(),
-                        ) {
-                            onPlaceClick(entity, id, getNameWithDisambiguation())
-                        }
-                    },
-                    onDelete = {
-                        onDeleteFromCollection(listItemModel.id, listItemModel.name)
-                    }
-                )
-            }
-
-            else -> {
-                // Do nothing.
-            }
-        }
-    }
-}
