--- app/src/main/java/ly/david/mbjc/ui/collections/recordings/RecordingsByCollectionScreen.kt
@@ -1,75 +1,0 @@
-package ly.david.mbjc.ui.collections.recordings
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.collectAsLazyPagingItems
-import ly.david.data.domain.listitem.RecordingListItemModel
-import ly.david.data.getNameWithDisambiguation
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.ui.common.paging.PagingLoadingAndErrorHandler
-import ly.david.ui.common.recording.RecordingListItem
-import ly.david.ui.common.rememberFlowWithLifecycleStarted
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-internal fun RecordingsByCollectionScreen(
-    collectionId: String,
-    isRemote: Boolean,
-    filterText: String,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    onRecordingClick: (entity: MusicBrainzEntity, String, String) -> Unit = { _, _, _ -> },
-    onDeleteFromCollection: (entityId: String, name: String) -> Unit = { _, _ -> },
-    viewModel: RecordingsByCollectionViewModel = hiltViewModel(),
-) {
-    val entity = MusicBrainzEntity.RECORDING
-    val lazyListState = rememberLazyListState()
-    val lazyPagingItems: LazyPagingItems<RecordingListItemModel> =
-        rememberFlowWithLifecycleStarted(viewModel.pagedEntities)
-            .collectAsLazyPagingItems()
-
-    LaunchedEffect(key1 = collectionId) {
-        viewModel.setRemote(isRemote)
-        viewModel.loadPagedEntities(collectionId)
-    }
-
-    LaunchedEffect(key1 = filterText) {
-        viewModel.updateQuery(filterText)
-    }
-
-    PagingLoadingAndErrorHandler(
-        modifier = modifier,
-        lazyListState = lazyListState,
-        lazyPagingItems = lazyPagingItems,
-        snackbarHostState = snackbarHostState
-    ) { listItemModel: RecordingListItemModel? ->
-        when (listItemModel) {
-            is RecordingListItemModel -> {
-                SwipeToDeleteListItem(
-                    content = {
-                        RecordingListItem(
-                            recording = listItemModel,
-                            modifier = Modifier.animateItemPlacement(),
-                        ) {
-                            onRecordingClick(entity, id, getNameWithDisambiguation())
-                        }
-                    },
-                    onDelete = {
-                        onDeleteFromCollection(listItemModel.id, listItemModel.name)
-                    }
-                )
-            }
-
-            else -> {
-                // Do nothing.
-            }
-        }
-    }
-}
