--- app/src/main/java/ly/david/mbjc/ui/collections/releasegroups/ReleaseGroupsByCollectionViewModel.kt
@@ -1,67 +1,0 @@
-package ly.david.mbjc.ui.collections.releasegroups
-
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-import ly.david.data.musicbrainz.MusicBrainzAuthState
-import ly.david.data.musicbrainz.getBearerToken
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.data.network.ReleaseGroupMusicBrainzModel
-import ly.david.data.network.api.BrowseReleaseGroupsResponse
-import ly.david.data.network.api.MusicBrainzApiService
-import ly.david.data.room.collection.CollectionEntityDao
-import ly.david.data.room.collection.CollectionEntityRoomModel
-import ly.david.data.room.relation.RelationDao
-import ly.david.data.room.releasegroup.ReleaseGroupDao
-import ly.david.ui.common.releasegroup.ReleaseGroupsByEntityViewModel
-import ly.david.ui.common.releasegroup.ReleaseGroupsPagedList
-
-@HiltViewModel
-internal class ReleaseGroupsByCollectionViewModel @Inject constructor(
-    private val musicBrainzApiService: MusicBrainzApiService,
-    private val collectionEntityDao: CollectionEntityDao,
-    private val relationDao: RelationDao,
-    releaseGroupDao: ReleaseGroupDao,
-    releaseGroupsPagedList: ReleaseGroupsPagedList,
-    private val musicBrainzAuthState: MusicBrainzAuthState,
-) : ReleaseGroupsByEntityViewModel(
-    relationDao = relationDao,
-    releaseGroupDao = releaseGroupDao,
-    releaseGroupsPagedList = releaseGroupsPagedList,
-) {
-
-    override suspend fun browseReleaseGroupsByEntity(entityId: String, offset: Int): BrowseReleaseGroupsResponse {
-        return musicBrainzApiService.browseReleaseGroupsByCollection(
-            bearerToken = musicBrainzAuthState.getBearerToken(),
-            collectionId = entityId,
-            offset = offset
-        )
-    }
-
-    override suspend fun insertAllLinkingModels(
-        entityId: String,
-        releaseGroupMusicBrainzModels: List<ReleaseGroupMusicBrainzModel>,
-    ) {
-        collectionEntityDao.insertAll(
-            releaseGroupMusicBrainzModels.map { releaseGroup ->
-                CollectionEntityRoomModel(
-                    id = entityId,
-                    entityId = releaseGroup.id
-                )
-            }
-        )
-    }
-
-    override suspend fun deleteLinkedEntitiesByEntity(entityId: String) {
-        collectionEntityDao.withTransaction {
-            collectionEntityDao.deleteAllFromCollection(entityId)
-            relationDao.deleteBrowseEntityCountByEntity(entityId, MusicBrainzEntity.RELEASE_GROUP)
-        }
-    }
-
-    override fun getLinkedEntitiesPagingSource(entityId: String, query: String, sorted: Boolean) =
-        collectionEntityDao.getReleaseGroupsByCollection(
-            collectionId = entityId,
-            query = "%$query%",
-            sorted = sorted
-        )
-}
