--- app/src/main/java/ly/david/mbjc/ui/collections/works/WorksByCollectionScreen.kt
@@ -1,75 +1,0 @@
-package ly.david.mbjc.ui.collections.works
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.collectAsLazyPagingItems
-import ly.david.data.domain.listitem.WorkListItemModel
-import ly.david.data.getNameWithDisambiguation
-import ly.david.data.network.MusicBrainzEntity
-import ly.david.ui.common.listitem.SwipeToDeleteListItem
-import ly.david.ui.common.paging.PagingLoadingAndErrorHandler
-import ly.david.ui.common.rememberFlowWithLifecycleStarted
-import ly.david.ui.common.work.WorkListItem
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-internal fun WorksByCollectionScreen(
-    collectionId: String,
-    isRemote: Boolean,
-    filterText: String,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-    onWorkClick: (entity: MusicBrainzEntity, String, String) -> Unit = { _, _, _ -> },
-    onDeleteFromCollection: (entityId: String, name: String) -> Unit = { _, _ -> },
-    viewModel: WorksByCollectionViewModel = hiltViewModel(),
-) {
-    val entity = MusicBrainzEntity.WORK
-    val lazyListState = rememberLazyListState()
-    val lazyPagingItems: LazyPagingItems<WorkListItemModel> =
-        rememberFlowWithLifecycleStarted(viewModel.pagedEntities)
-            .collectAsLazyPagingItems()
-
-    LaunchedEffect(key1 = collectionId) {
-        viewModel.setRemote(isRemote)
-        viewModel.loadPagedEntities(collectionId)
-    }
-
-    LaunchedEffect(key1 = filterText) {
-        viewModel.updateQuery(filterText)
-    }
-
-    PagingLoadingAndErrorHandler(
-        modifier = modifier,
-        lazyListState = lazyListState,
-        lazyPagingItems = lazyPagingItems,
-        snackbarHostState = snackbarHostState
-    ) { listItemModel: WorkListItemModel? ->
-        when (listItemModel) {
-            is WorkListItemModel -> {
-                SwipeToDeleteListItem(
-                    content = {
-                        WorkListItem(
-                            work = listItemModel,
-                            modifier = Modifier.animateItemPlacement(),
-                        ) {
-                            onWorkClick(entity, id, getNameWithDisambiguation())
-                        }
-                    },
-                    onDelete = {
-                        onDeleteFromCollection(listItemModel.id, listItemModel.name)
-                    }
-                )
-            }
-
-            else -> {
-                // Do nothing.
-            }
-        }
-    }
-}
